Compile `user_dir` for all Unix & Redox targets
This change broadens the list of targets which will conditionally
compile the `user_dir` and `user_dirs` functions. Prior to this change,
only the Linux and Redox OS targets will compile an implementation
meaning that other Unix derivatives such as Android, FreeBSD, NetBSD,
Darwin, etc. would not have a compiled implementation of these
functions. Once paired and consumed by the upstream `dirs` crate this
would lead to a compilation failure on these "extra" targets. This
change allows all Unix family targets in addition to the Redox OS target
to compile the same implementation.

Closes #1

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>